split newly-massive 'pages' modules into UI portion, callback portion, and dependency service function portion
